Write Height, isBalanced, rebalance, driver functions
Wrote a Height funtion for finding the height of a specific node. Wrote
an isBalanced function to check whether the binary tree is balanced or
not. Wrote a rebalance function to create a new balanced binary tree
from the nodes in the current (unbalanced) binary tree. Wrote a driver
functions that uses most of the written modules, and manipulates the
binary trees as specified. Cleaned the app file so that all functions
other than the driver function are imported from seperate modules,
including the buildTree function that consumed the most space previously.